requirements-ci.txt
# Minimal dependencies for GitHub Actions unit tests.
# Heavy ML dependencies from requirements.txt are intentionally excluded.
fastapi
pydantic-settings
aio-pika
SQLAlchemy
psycopg[binary]
pytest
httpx
